Skip typing.NoReturn tests on Python < 3.6.2
typing.NoReturn was introduced in Python 3.6.2. Move the tests for
issue #4122 to a separate file and skip when Python is too old.
