+++ tests/functional/i/inconsistent/inconsistent_returns_noreturn.rc
@@ -1,0 +1,5 @@
+[testoptions]
+min_pyver=3.6.2
+
+[REFACTORING]
+never-returning-functions=sys.exit,sys.getdefaultencoding
