+++ tests/functional/i/inconsistent/inconsistent_returns_noreturn.txt
@@ -1,0 +1,1 @@
+inconsistent-return-statements:32:0:bug_pylint_4122_wrong:Either all return statements in a function should return an expression, or none of them should.
